fix for findStoreEntry() returning wrong entry in some cases
This would only surface when using DataStore::getRelat* with a
RelationsObject (instead of calling RelationsObject::getRelat*)
with initialised m_cache* members and with multiple StoreArrays
of the same class.

Only seen when looking at the new RecoTracks in the display's
DataStore browser. Standard simulation+reconstruction doesn't
seem to be affected (tested on a few thousand events)

Thanks to Nils for noticing.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@26494 b38c6c8f-bee8-465d-8448-9abe6915d496