Add possibility to create Geometry from database
This commits adds the possibility to construct the Geometry from the
Database. It adds classes to save the overal geometry configuration to
the db and extends the GeometryManager and Materials interface to either
use the previous Gearbox approach or to load a previously saved
configuration.

Everything should be backwards compatible for the moment. The Geometry
module has now new parameters:

  - useDB to load the geometry from the database
  - createPayloads to convert Gearbox to database payloads
  - payloadsIov to set the iov when converting Gearbox parameters

If useDB is set the Geometry module will load the list of materials and
sub detectors from the DB. It will create all Materials from the DB and
try to construct all sub detectors from DB first but fall back to
Gearbox parameters for the subdetectors which do not implement creation
from DB.

For this to work the CreatorBase interface has been extended by two member
functions: createFromDB() and createPayloads(). All sub-detectors should
convert their Geometry Creators to implement createFromDB() and
createPayloads(). The way to do this is to split the create() member
into two functions. The first should convert the Geabox parameters into
Payload objects and the second should create the actual Geometry from
these objects. Please see geometry/examples/DBCreatorExample.cc for
details.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@26621 b38c6c8f-bee8-465d-8448-9abe6915d496