parallel processing: fix problem introduced by loading first event early
If one of the processes did not get any event (more processes than events,
bad timing, etc.) then it would no process the first event of the job
and pass it to the output which leads to event duplications.

Fix that and adapt the test to be more sensible to this issue.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@26669 b38c6c8f-bee8-465d-8448-9abe6915d496