Database: Update unit tests
As intra run dependency is already setup correctly on updateEntry() we
won't get a callback when updateEvent() is called the first time.

TODO: Maybe we should callbacks as soon as they are added the first time
if the object is valid? As least as an option. Otherwise people have to
do an

<pre>
dbobject.addCallback(myfunction);
if(dbobject) myfunction();
</pre>

when initialization is performed in myfunction()

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@26865 b38c6c8f-bee8-465d-8448-9abe6915d496