Fix missing python traceback when redirecting
When running basf2 and redirecting the output (for example
"steeringfile.py &> output.log") it seems to happen that python
traceback for unhandled exceptions is not printed since the buffers are
not flushed. This is especially annoying in the validation when the log
file does not show any error at all.

running Py_Finalize() after PyErr_Print() seems to fix this.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@27169 b38c6c8f-bee8-465d-8448-9abe6915d496