geometry: Close geometry after creating it
This is needed for navigating the geometry. It's usually done by either
FullSim or the Ext module but if both are missing the geometry was not
closed properly.

This might cause some delay in startup since the G4RunManager used by
FullSim and Ext insists on closing the geometry again so we will
caluclate the optimizations twice but there is no good way around this
for now.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@27566 b38c6c8f-bee8-465d-8448-9abe6915d496