Introducing streamers to low level cdc tracking objects.
This softens the issue of transfering tracking objects at interprocess boundaries.
As it turn out ROOT is able to establish and propagate pointers between objects
which presented the hardest conceptual problem in this endeavour.
The remaining obstacle was the link between CDCWireHit and CDCWire objects since the
wires are never placed in the DataStore. Instead we now add the WireID as a member
to the CDCWireHit for streaming and reset the wire based on that id on unpacking.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@28399 b38c6c8f-bee8-465d-8448-9abe6915d496