Expose more TTree write options as parameters
to allow studies if reading parts of the events over xrootd can be
improved.

New Parameters:
   - autoFlushSize: size in events (>0) or bytes (<0) after which to
     flush all baskets, smaller values should improve event locality in
     the file but hurt sequential performance
   - autoSaveSize: size in events (>0) or bytes (<0) after which to
     flush tree metadata to the file to allow recovery in case of crash
   - basketSize: basket size in bytes, smaller baskets will increase
     event locality but hurt sequential performance

Together with the splitLevel parameter this exposes all input/output
related optimization parameters I could find. We might optimize the
basketesizes for different branches by hand but that sounds like a a lot
of fiddling

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@28683 b38c6c8f-bee8-465d-8448-9abe6915d496