Remove obsolete FIXME indication.
The MCMatcherTracks is insensitive to missing hits StoreArrays and would work just fine
without the explicit indication of the hits to be used.
However I am leaving the explicit setting in for the common scenario of simulating with all detector components
and execute tracking on only a part of the tracking detector.

In this way the matching clips to the components mentioned in the reconstruction and
not to the available hits StoreArrays from the components used in the simulation, which would otherwise
give a false impression of the performance of the limited track finder in relation to the whole recorded track in the full detector.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@28719 b38c6c8f-bee8-465d-8448-9abe6915d496