Initial commit of new mva package:
1. It implements a common interface for training and applying MVA frameworks
in BASF2 like FastBDT, NeuroBayes, TMVA, and FANN, as well as Python-based
frameworks like SKLearn, XGBoost and Tensorflow.

2. Training can be done using C++, Python or Shell Commands

3. The weightfiles of the methods can be saved on disk in a ROOT
file or in the BASF2 condition database. They are run-dependend.

4. The training can be applied using C++, Python, Shell Commands or the
MVAExpert module.

5. The mva package deprecates the TMVA Interface and TMVAExpert module
in the analysis package. All algorithms FEI, FlavourTagging,
ContinuumSuppresion should migrate to the new mva package instead.

6. Support for data-driven SPlot and boosted SPlot trainings is
implemented

7. We want to add plotting and evaluation functionality to the package,
especially the code which is already present in the framework (b2plot)
and tracking package.

8. Other packages like tracking and ecl are welcome to migrate/replace their
existing mva infrastructure to/with the common mva package.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@28957 b38c6c8f-bee8-465d-8448-9abe6915d496