Make the framework compile without undefined symbols
This commit changes the linking options to not allow any undefined
symbols in the produced binaries or shared libraries. This should make
the "missing libraries" check in the build bot superfluous since we are
guaranteed that all symbols are defined.

It will also set the "--as-needed" flag which will reduce the amount of
emitted dependencies as it will silently drop libraries which are
specified as dependencies but not actually required.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@29355 b38c6c8f-bee8-465d-8448-9abe6915d496