Add library dependency check to scons
Now there are two more options to scons: --check-extra-libraries and
--check-missing-libraries. After the build this will check all created
.so files and compare the list of libraries specified in the SConscript
to the list of libraries which are acturally linked in. It will report
unneeded dependencies and missing direct dependencies.

For now it will *only* consider internal libraries, that is ones located
in $LIBDR due rather complicated name matching if so versions are
included.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@29356 b38c6c8f-bee8-465d-8448-9abe6915d496