PXDClusterizer: make sure there are no segfaults for invalid Cell Ids
the pxd clusterizer keeps a list of clusters in the last and the current
row and the length (aka number of columns) needs to be known.

This patch adds a check to make sure only pixels within these columns
are considered, if u cell id <0 or > maxU an std::out_of_range exception
will be thrown which is then converted to a warning message that this
pixel is ignored for clustering.

There's also a test for this behaviour and the number of columns is set
to 250.

git-svn-id: https://belle2.cc.kek.jp/svn/trunk/software@29549 b38c6c8f-bee8-465d-8448-9abe6915d496