Remove prefix when checking library dependencies
SCons seems to remove the 'lib' prefix automatically when specifying
libraries to link against, e.g. `LIBS=['libMinuit']` results in `-lMinuit`.

This is all fine but the --check-{extra,missing}-libraries option we
have did not take this into account. Now it will strip the 'lib' prefix
and emit a LIB_WARNING message.

Closes BII-1668