Make size of simulation top volume dynamic
This commit adds the feature that the size of the simlation top volume
will be dynamically enlarged to fit all placed volumes. If the size in
any direction has been configured as zero or negative it will be
silently enlarged, if it had been set to a positive value which is not
big enough a warning will be emitted.

The volume will never be smaller then the configured value. This new
feature should not have any impact on normal users, it only applies if any of
width, length or height are missing from the Global configuration in the
xml or set to <=0 or if the volume is not large enough (for example when
adding FarBeamLine component).

This also fixes BII-1052 as it tries to make the documentation more
clear and point out that all dimensions are half lengths.