Add CrashHandler for graceful handling of crashes in HLT
Module takes a Path to execute in a crash-safe manner. For events where
any of the contained modules crash or produce FATALs, no output from them
will be available in the DataStore, and the return status of the module
will be set to 0/false.

This is meant for the HLT, where interruption of data-taking because of
individual events is unacceptable.  There will be hell to pay if I catch
anyone using this instead of fixing underlying issues.

Closes BII-1666