Remove excessive Warnings about duplicate payloads
The idea is that there should only be one payload per key for a given
run. However since we are currently still in development and deletion of
iovs seems to not be working at the moment we will have multiple
payloads for  the time being. The conditions interface will just pick
the one with the highest revision so this should be fine.

This commmit thus replaces the warnings printed for each duplicate
payload with debug messages and will just print a list of the payloads
which were present more then once as INFO message after all of the
payload information is parsed.