Changed order of SpacePoints contained in SpacePointTrackCandidates. from Outermost -> Innermost to Innermost -> Outermost
MomentumSeedRetriever:
Since the old trackletfilter class still requires the
outermost->innermost order the loop is reversed.
The positionSeed needs to be taken from the innermost hit. Therefore it
is taken from the first spacePoint now instead of the last one.

Additionally the method to retrieve the magnetic field was changed, since
the old method returned the wrong value.