Updating the sha3 hash algorithm to the current version of the reference implementation.
As before the relevant sources are taken from https://github.com/gvanas/KeccakCodePackage,
only modified in whitespace to be congruent with the basf2 style guide.

Since the last update the code structure and file names have been changed profoundly.
To guide future updates an update-hints.sh files was included containing a rough outline what has been done in this update.

Also some preprocessor macros needed to be included in the framework/SConscript to limit the varity of algorithms to the ones we need.