Wrap things in Belle2 namespace
The intel compiler brings it's own cosd which is incompatible with the
*globally* defined cosd in shapes.h. To make all this more robust I
wrapped shapes.h and BelleCrystal.h in the Belle2 namespace and moved
some global variables in barrel.cc in a anonymous namespace.

The diff is large due to fixstyle but a `git diff -b` to ignore
whitespace changes shows that the changes are indeed rather simple.