Recreate digits.root in rawdata/tests
The digits.root used by the rawdata/tests/packers.py was outdated and
caused the test to fail. Since there was no instruction on creating this
file I expanded the packers.py to create it automatically if it is
missing.

In futute it should be enough to delete digits.root and run the test to
create a new version. Closes BII-1823.