Modify download and lookup scheme (BII-1216, BII-1663)
Instead of one base url and one local directory like before we now have
a list of local directories to look before downloading and use the
baseUrl included in v2 of the api to find the payload on the server.
(For v1 access we take a hardcoded base url)

So now we

- look in all defined local directories for the file. If we find one
  with the correct name we do the checksum and return if it matches.
- if we couldn't find any file we download from the central server and
  try to put it in the first defined local directory.

I removed the read lock as I don't see how this could be a problem: If
another process is already downloading the checksum test will fail and
we will obtain a write lock and recheck the checksum before downloading.
This makes readonly directories independent of fnctl support.

In addition the local directories support different layout structures:
One flat directory like we used before, following the logical hirerachy
defined by the server or using a git-style checksum-subdirectory scheme
where payload foo/bar_v1.root with checksum 0123456789ABCDEF will be
put into 0/12/bar_v1.root in the local directory. The default local
directory is still a flat centraldb directory in cwd.