Make sure Database instance is cleaned up
Database singleton was not cleaned up correctly at the end because it
was a static Database*; This commit changes this to a
std::unique_ptr<Database> so unless aborted or killed otherwise the
destructor of the Database is now called.

This is needed to clean up potential temporary files created when
downloading payload files.