Changed isNotContinuumEvent var to handle all Upsilon(NS), isContinuumEvent calls isNotContinuumEvent.
Updated isNotContinuumEvent function to only check the MCParticle in the
event with index=1 (i.e. initial particle) and return true if it is any
of the Upsilon(NS), N = 1 - 6 resonances according to PDG code.

Also changed isContinuumEvent to simply return opposite of
isNotContinuumEvent, rather than repeating the code for the entire
function with opposite return values as was done before.