Set isContinuum vars in EventVariables to check only first element of MCParticle array, also handles non-existent MCParticle lists.
Added suggestions in Pull requests #109:
isContinuumEvent passes nullptr when calling isNotContinuumEvent to make
obvious that this is an event-level variable.

isNotContinuumEvent now checks only the first particle in the MCParticle
list and that it's an Upsilon(NS) with no parent, instead of looping over entire list.
Also handles when no MCParticle list is found throwing B2ERROR.