Add option to generate a zr fieldmap averaged over phi
The CreateFieldMap module now has a scan type "zr" which will obtain the
magnetic field at different phi positions and average between them and
create a ROOT file which contains the histograms for |B|, Bz and Br in
the zr plane.

This also changes the output of the other scans to |B|, Bx, By, Bz
histograms.