Handle BackgroundInfo when merging files (BII-1930)
This adds special handling when merging files containing background
information:

- check that the BackgroundInfo objects are present in all or no files
- check that they are all identical except for the reused counter
- sum up the reused counter and increment by (n-1) where n is the number of files

All the code added to the merge tool by this commit should be removed and the
BackgroundInfo class should implement the merging on its own and inherit from
Mergeable (raising an exception if some aspects cannot be merged). These are
minimal invasive changes to be included in release-00-08-00.