Relax BackgroundInfo checks
to allow for different background file samples to be used the
compatibility check is now more relaxed. We only check for

* same method, detector components, time windows, wrap around and max
  ECL energy deposition.
* same list of background types
* same background scaling factors

Combined with the overal checks on same release/steering file and so
forth this should be strict enough.

The BackgroundInfo is now no longer merged but all of the BackgroundInfo
instances are added to the output file to keep all information.