Add timeout option to test_basf2 (BII-1951)
This adds a "-t|--timeout=SECONDS" option to test_basf2 which allows
limiting the allowed time for each test case. If the timeout is reached
the test will be killed and marked as failed.

The default is to have no timeout which is compatible with what we had
so far.