Fix duplicate creation of sub components in SVD
This probably could apply to PXD as well but seems to not occure there.
It now works and reprouces the exact geometry with or without database
except for one small rounding difference due to double precision vs.
decimal text value.

The underlying problem that information for components is present more
than once in the database is not fixed.