don't mix runs in parallel processing
When a new run starts, the input process will wait for the attached
buffer in Tx::beginRun() until it is empty, and all processes reading
from it are 'not busy', i.e. have finished processing the events
from the previous run.

Adds a new field 'nbusy' to the RingBuffer metadata, can be checked
using monitor_ringbuffers.