Remove assumption that TObject is first base class
If we have a class with multiple inheritance where TObject is not the
first base class the pointer to TObject starts not at the beginning of
the object so a reinterpret cast or cast from void* will destroy it.

So use ROOT to obtain the correct offset when creating objects.