Fix multiple inheritance in MsgHandler
TBuffer::ReadObjectAny() can handle multiple inheritance objects with
TObject not being first baste correctly but we used it wrong: It needs
to get the class the pointer will be, not the class the object is.

This means we can also omit reading the class from the buffer twice.
This was not working correctly anyway as it did not start at the
necessary offset.

Also make error checks more robust.