Properly initialize EvtHeader structure
up to now the EvtMessage class did not properly initialize all members
of the EvtHeader structure at the beginning of an EvtMessage. The
reserved fields were just taken from uninitialized memory.

We now fix this by

 * adding a default value for all fields
 * using placement new to initialize the memory area according to these
   values
 * add a version information at the end of EvtHeader which starts with
   0xBEEFED + 8byte version information to check whether the header has
   the fixed initialization or not.
 * return 0 for flags if the version of the header is not valid as the
   memory might be uninitialized in this case.

This version field has the additional benefit that EvtMessages could now
be found more easily in the binary stream in case a size field gets
corrupted.