Enable some more clang warnings
Clang has a very nice set of diagnostics. Two I would like to enable are
 * -Wdocumentation: warn if doxygen documentation seems to be wrong:
   parameters which cannot be found in the function, @return in void
   functions, empty @return or @param statements
 * -Wheader-hygiene: warn about `using namespace` in the top level of
   included files.

The first one is nice to have but I could live without it. The second
should really be enabled as this can lead to all kinds of trouble.

This commit also disables all warnings when compiling dictionaries:
usually this should only be warnings which were already seen in another
translation unit using the classes or are related to ROOT dictionary
creation hacks (`namespace std {}; using namespace std;`).