Allow to intercept output which doesn't use logging facilities.
This adds a new namespace containing a few classes to intercept output
to stdout or stderr. It does this by modifying the underlying file
descriptors so it works with C++, C and fortran code and can be used to
silence output we don't have control over or to capture output and
convert it to log messages.