Add SkimName to FileMetaData (BII-2142)
This modifies skimOutputUdst() to add the skim name to the metadata of
the output file.

The first argument is now the skim name which will be added to the
metadata as a data description "SkimName". There's now a new
keyword-only argument "outputFile" which allows to specify the output
filename.

So a previous call

    skimOutputUdst("output/MySkim", particleList)

should now be converted to

    skimOutputUdst("MySkim", particleList, outputFile="output/MySkim")

otherwise the skim name will be interpreted as "output/MySkim".

To access the name of the skim from the output file the best way is to
parse the json output of showmetadata:

    showmetadata --json output/MySkim.udst.root

For example in python this could be done using

    import json
    import subprocess

    data = json.loads(subprocess.check_output(["showmetadata" , "--json", "output/MySkim.udst.root"]).decode())
    print(data["dataDescription"]["SkimName"])