Minimize libGLU/TEve dependency (BII-2158)
libgenfit2 usually also contains the parts which are responsible for the
EventDisplay and thus links against TEve and libGLU which might not be
installed everywhere. As libgenfit2 is used widely this makes almost all
of our code depend on TEve indirectly if the Belle2 Software was
compiled with TEve support which is problematic for grid jobs.

This commit splits genfit into two libraries: libgenfit2 and
libgenfit2_eventDisplay which contains the eventDisplay part
(obviously).

Before the patch a lot of libraries/modules pulled in libGLU which can
be checked by executing the following line in the libraries/modules
directory:

    for file in *.so; do ldd $file | grep GLU && echo $file; done

After this commits this only affects the display module and the
genfitVisModule so almost all scripts will run unaffected even if libGLU
is not installed on a particular system.