Fix intra run dependencies with sub event paths (BII-2256)
When combining intra run dependencies with sub event paths
(`path.for_each()`) the intra run dependency was not working anymore
because the list of existing intra run dependencies was cleared when
updating the DBStore even if the payload remaind valid. As
SubEventModule also calls DBStore::update() the intra run dependency
handling was lost.

This commit fixes that by moving the intraRunDependencies from a vector
to a set so that we can efficiently remove only payloads which became
invalid.