Allow overriding the fallback database or global tag from environment variables
This commit introduces two environment variables "BELLE2_CONDB_FALLBACK"
and "BELLE2_CONDB_GLOBALTAG" which can be used to specify which file to
use as fallback database and which global tag to use by default

These environment variables only change the default setting and can be
overriden in the steering file by explicitely configuring the database.

If an empty value is set for these variables (in contrast to not having
the variable at all) the fallback database or access to the central
database will be disabled.

More then one global tag can be specified by separating them with
whitespace.