test_basf2: extend XML output
this commit adds stdout/stderr of failing tests to the xml output which
should make it much easier to pin down problems for failing framework
test. It should have no change for normal usage and just adds either the
contents of stdout/stderr or the diff between the template and the
actual output to the body of the "failure" tag in the xml.