Add scope guard to IOIntercept to simplify use.
In case the io interception should be performend in a certain scope
which might have different exit paths (return, exception, break) the new
`IOIntercept::start_intercept()` will make sure that the intercept is
finished in all cases.

This is intended for use cases where the intercept class itself is a
member or otherwise longer lived so the automatic cleanup in the
destructor does not apply for the limited scopes.