Allow to configure naming scheme for seqroot (BII-2356)
This commit enables the SeqRootInput/Output module to configure the
naming scheme for the sroot files to be read/written. Now the filename
parameter can be a printf/boost::format pattern to specify how the name
should be generated. For example

   main.add_module("SeqRootOutput", outputFileName="output-part%05d.sroot",
                   fileNameIsPattern=True)

will create the sroot files starting at "output-part00000.sroot" and
incrementing the number for each file instead of appending it to the end
of the filename. By default the SeqRootOutput keeps the default
behaviour of just creating .sroot-N files.

The same thing applies to SeqRootInput: by default it will try to open
.sroot-N files when the current file is finished but using the
`fileNameIsPattern` argument one can tell it to read additional sroot
files with a different naming scheme.