Make sure the ParticleCombiner does not use infinite memory (BII-2357)
Without a proper limit the `ParticleCombiner` will happily combine as many
candidates as it can find and this is an factorial combinatorial
problem this can result in infinite memory consumption. (recently double
charm skimming jobs were killed with 15-20GB ram consumption)

On the other hand lists with that many candidates seem questionable
so it seems reasonable to introduce a default limit on the number of
candidates to make sure memory consumption does not grow without limit.

This commit adds a default limit of 2000 candidates and print a
warning if this limit is exceeded. It will also exposes the limit in the
`modularAnalysis.py` so that users can easily increase the limit for
their special use case.