Allow downloading of multiple global tags (BII-2348)
There are two new parameters to `conditionsdb download`: `--tag-pattern`
and `--tag-regex`. Those two parameters are mutually exclusive.

If `--tag-pattern` is given the `TAGNAME` is interpreted as a shell
pattern for the tagname and all global tags matching this pattern will
be downloaded.

If `--tag-regex` is given the `TAGNAME` is interpreted as a python
regular expression for the tagname and all matching global tags will be
downloaded

If `-c` is given as well the generated database files will be called
`DIR/TAGNAME.txt`

So for example to download all tags by the Global tag coordinator in the
into the directory `outputdir` we can now call

    conditionsdb download -c --tag-pattern 'GT_*' outputdir/