Consolidate module folders
Splitting all the modules in different shared libraries just hurts
startup performance but does not gain anything except me being confused
where all the modules are.

Now we split them by dependency and/or deprecateability: Gearbox is
slashed to not be used by default in the future so we keep it
separately.

Also: remove fastseqrootinputModule. This was not working for quite some
time now.