Fix radiation sensors for SVD in phase2
The xml does not contain radiation sensors for VXD in phase2 so the code
correctly did not fill the parameters. But by default the parameter
object returns a empty instance of the radiation parameters and no check
was performed on this before creating the diamonds itself

This should be fixed also for PXD and maybe a bit more explicit then
checking the sub detector name but it should work for the release.