Changed the logics in the STM.
The trigger menu should now be downloaded directly from the DB (not implemented in the moment),
so there is no need to add a parameter for it. Same thing for the acceptOverridesReject parameter.

Additionally, the calculation of the total result is now included in the module and not
in the software trigger result class.