Fix order when downloading a global tag (BII-2438)
In case of duplicate payloads there are two slightly incompatible
procedures:

 - the CentralDatabase will use the payload with the highest revision
 - The LocalDatabase will use the last payload in the file which maches
   the given iov.

To make the `conditionsdb download` work consistently we have to order
the created database.txt file correctly. So far this was done slightly
incorrect by ordering the lines as strings which leads to problems when
the revisions become bigger than 9.

This patch sorts the lines correctly by treating revisions and exp/run
numbers as integers while sorting.