Remove IOIntercept from ParticleVertexFitter (BII-2453)
The IOIntercept is only used for Rave vertex fits but the intercept is
also present in RaveKinematicVertexFitter. At the moment this leads to 2
open file descriptors per ParticleVertexFitter module open during the
runtime of the program.

Also make sure the IOIntercept is used correctly in raveInterface using
a RAII guard and optimize by using a single static instance of the
output interceptor which is only created if Rave is used.