Renamed the pre_general_cosmics_run and add a top_in_counter
The new data_taking_period is now also used to describe the gcr2017 setup,
which is the default. Also, the top_in_counter can now be configured
independently (is it possible to know from the data_taking_period, when to turn it on?)
It is set to False as default.
Additionally, the time extraction is turned on by default, but still using the -70, to range.