Use the new naming schema also during generation.
Also, removed the check for the right geometry (as this is now not that simple
anymore) and added a check to not use 201607, as this period is somewhat special.

Still, I am not sure when to add the cosmics selector. Now, I am adding it, if
the data taking period is not in normal or gcr2017.