Correct check for usability of TEve
Up to now we checked whether the libEve.so library is present in the
root directory and can be used and that libGLU.so is present.

The latter is only available if libglu-dev is installed, even if libglu
is present but we do not need it for compilation. We only need to make
sure that we can link against TEve.

So use the correct CheckLibWithHeader call to make sure we can
instantiate a TEveViewer().