Fixing bugs and add shared ptr to CalibrationAlgorithm
* Unfortuantely this commit contains two changes
* One is a reworking of how previously created and saved objects
are found by the collector/objmanager. And then new objects are
created. It should fix a bug where returning to a previous exprun
causes directories to get screwed up and FindObject/CloneTree to segfault
* The other change is to CalibrationAlgorithm. It now returns a shared_ptr
to the created/merged calibration objects.
* This allows us to hold a map of the object name+runrange -> object ptr so
that we only merge ONCE on the first getObjectPtr() call. And subsequent calls
look up the map.
* The map is cleared at the beginning of a new execute() call which deletes the objects.
But shared_ptr potentially allows a caller to save a reference to the shared_ptr between calls
and so prevent previously merged objects from being deleted. Seems unlikely to be useful
but it's there.