Fixed segfault and removed the CalibRootObj intermediate objects
* Segfault was due to the registered template objects in the manager
not being deleted before the end of the python process (not the end
of the basf2 process() call, but living beyond even that).
* Apparently the destructor of modules is never called? Hence the issue.
* Added a manual cleanup in the terminate() function and had to enforce
terminate in all processes to make sure that each parallel process cleans up
itself.
* Also removed the CalibRootObj+Base classes from the manager. Now that we
have a better undersanding of the necessary methods to call on each object we
can get away with just using mergeable objects ultimately inheriting from TNamed.
So no need for the template/base class shenanigans. Will remove the files soon.