Don't select hit from those TS hits not related to the early cluster.
TS selection process was originally independent to clustering and peak finding,
and it used all TS hits, even those not contributing to the Hough cluster.
Now they are checked before being selected and related to the track output,
so that an early cluster will not be related to the TS hits ahead in time.