Change fallback database
This changes the fallback database to look into cvmfs by default which
will enable to run without access to the global database. The payloads
in framework/data are not sufficient for that case.

Users without cvmfs need to download the global tag

    conditionsdb download -c GT_gen_prod_003.19_release-00-09-02 snapshot_dir

and set BELLE2_CONDB_FALLBACK to the created database.txt tag.