Always prefer cvmfs payloads over downloading
This commit adds the cvmfs conditions directory as search directory for
ConditionsDatabase access. This means that the conditions database
client will look into `/cvmfs/belle.cern.ch/conditions/` to see if it
can find the payloads there before downloading them.

This should be especially useful for FEI and the Magnetic field (44MB
payload) as those will just be obtained by cvmfs. The downside is that
this will probably remove most load from the conditions db file server
and put it on cvmfs instead but it will greatly help with scalability
and should allow running FEI on the grid without disabling access to the
central database.