Allow to set a specific proxy for ConditionsDB access
This introduces `$BELLE2_CONDB_PROXY:

* if not set use default proxy configuration (e.g. honor `$http_proxy`)
* if set to "" use direct connection
* otherwise handed directly to curl as proxy configuration which as of
  now is defined as below

---

Set the proxy to use for the upcoming request. The parameter should bae
astring holding the host name or dotted numerical IP address. A
numerical IPv6 address must be written within [brackets].

To specify port number in this string, append :[port] to the end of the
host name. If not specified, libcurl will default to using port 1080 for
proxies.

The proxy string should be prefixed with [scheme]:// to specify which
kind of proxy is used.

- **http://**
  HTTP Proxy. Default when no scheme or proxy type is specified.
- **https://**
  HTTPS Proxy. (Added in 7.52.0 for OpenSSL, GnuTLS and NSS)
- **socks4://**
  SOCKS4 Proxy.
- **socks4a://**
  SOCKS4a Proxy. Proxy resolves URL hostname.
- **socks5://**
  SOCKS5 Proxy.
- **socks5h://**
  SOCKS5 Proxy. Proxy resolves URL hostname.