Moving over to runners/strategies
This adds in the ability to change the way the CAF processes a calibration
by changing the class used to run a particular stage.
It introduces Strategies that run a single algorithm over an IoV and runners
which run multiple Algorithm Strategies.
It also makes the Calibration class inherit from a base class and Thread, so that
if someone wanted to they could implement their own Calibration class that doesn't
require the use of collecotr+algorithm. Although it would take a lot of work.