Provide phase2/phase3 geometry with one XML file
This commit adds the phase2 detectors to the Belle2.xml but disabled by
default. This allows the geometry to load the list of detector
components and all supported geometry components from the database and
fall back to XML definitions for the phase2 detectors and the beampipe.

It also updates the geometry/examples/createGeometryPayloadss.py to
creatcreate phase2 and phase3 geometry configuration for the database.