Use geometry from the Database for add_(simulation/reconstruction)
This changes add_simulation and add_reconstruction to create the
Geometry from Database configuration instead of XML files. The Geometry
module itself will still prefer XML to keep most of the validation/study
scripts working as before.

After the release I should add a warning to the geometry module if the
Geometry is not created from the Database.