Updating Millepede collector/algorithm
Changed the collector and algorithm to use "chi2_per_ndf" instead of
a variable name containing a '/' which messes up root directory creation.

Small change in Algorithm to test if a GBL data tree actually exists before
testing the number of entries. Maybe testing the entries is unneeded.

Updated the python Millepede class to use the new CAF with histomanager and
algorithm.setInputFileNames