Add PXDDigitSorter when using BGOverlay (BII-2770)
The background overlay adds digits at the end of the digit list but the
clustering expects them to be sorted. This can lead to process abort if
only the first sensor contains a signal digit but at the very least it
causes wrong clustering: The signal and background digits are clustered
separately.