Improve message for failed module creation
If an exception occured during module creation we silently dropped it
and showed some non-specific error message. This commits

* adds a custom python exception in case of the normal "Module could not
  be created" error (if the library could not be found or some such).
* split the exception handler to write a non-specific error message for
  the normal case (we know we created ERROR messages before) and a part
  which will show the e.what() for other exceptions

A nicer way might be to change the registerModule() function to raise
exceptions with a correct error message instead of the "Could not be
created: ModuleName" text, when we could just print e.what() regardless.

However then I have to check where we call these and make sure behavior
did not change. This is a safer approach as it still will print the same
warnings and errors, the returned exception type is different but still
inheriting from RuntimeError.