Add custom b2 domain to sphinx
This allows to document our modules and variables in sphinx in a
automatic way and adds directives to sphinx to show a list of modules or
variables:

.. b2-modules::
   :library: libcore.so
   :modules: EventInfoSetter, EventInfoGetter
   :regex-filter: .*
   :no-parameters:
   :io-plots:

.. b2-variables::
   :group: PID
   :variables: A, B, C
   :regex-filter: .*

They both have options to select the modules/variables to be displayed
or set some options but they can be called without any options to show
all known modules or variables.

This will also create custom indices for basf2 modules and basf2
variables.